+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Fix NameError on unknown line indicator.
+
 ## 0.3.0 - 2021-12-24
 
 ### Added